Update README.md
A graduate in Biology was having difficulty in finding a job. He saw an advert in one of the daily newspapers for a job at a zoo.🗞️👀

In the interview, the manager told him that their gorilla🦍, which had been a tourists attraction has died so they needed someone to dress up and pretend as a gorilla🦍. 

The graduate was embarra$$ed, but since the salary was okay, he accepted the job.

The first day, he put on the gorilla skin and entered the cage, he started jumping up and down, beating his chest and roared like a gorilla. 

The next day, he put on a gorilla skin and started moving around the zoo again and mistakenly entered another cage and found himself staring at a lion🦁. 

The lion r0ared and rushed towards him. The scared graduate quickly forgot that he is a g0rilla and started shouting like a human, 🗣️"Help! Help!" The lion leaped onto him, knocked him to the ground and whispered in his ear👂*Dennis*, it's me Mike, your course mate."
My brother, No job in this country, Infact that croc0dile🐊 in that water is not a crocodilie, it is John😂😂😂😂😂😂😂😂